+++ examples/simple/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies
+  - vite-plugin-vercel@5.0.4
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - vite-plugin-vercel@5.0.3
+
+## null
+
+### Patch Changes
+
 - vite-plugin-vercel@5.0.2
 
 ## null