+++ packages/vercel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # vite-plugin-vercel
 
+## 5.0.4
+
+### Patch Changes
+
+- Replace workspace-root by @manypkg/find-root
+  - @vite-plugin-vercel/vike@5.0.2
+
+## 5.0.3
+
+### Patch Changes
+
+- Avoid potential url and path module conflict
+  - @vite-plugin-vercel/vike@5.0.2
+
 ## 5.0.2
 
 ### Patch Changes